category with auth-admin

+++ app/Http/Controllers/CategoryController.php
@@ -2,66 +2,52 @@
 
 namespace App\Http\Controllers;
 
+include '/Users/antoncaus/Desktop/usoft/app/Support/helpers.php';
 use Illuminate\Http\Request;
 use App\Models\Category;
-
+use Tymon\JWTAuth\Facades\JWTAuth;
+use App\Models\User;
 
 class CategoryController extends Controller
 {
     public function index()
     {
         //get Category
-        return Category::all();
+        if(isAdmin(auth()->user()))
+            return Category::all();
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    
     public function store(Request $request)
     {
         //create a Category
-        return Category::create($request->all());
+        if(isAdmin(auth()->user()))
+            return Category::create($request->all());
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function show($id)
     {
         //show Category
         return Category::find($id);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function update(Request $request, $id)
     {
         //update Category
-        $Category = Category::find($id);
-        $Category->update($request->all());
-        return $Category;
+        if(isAdmin(auth()->user())) {
+            $Category = Category::find($id);
+            $Category->update($request->all());
+            return $Category;
+        }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
-        return Category::destroy($id);
+        if(isAdmin(auth()->user()))
+            return Category::destroy($id);
         //delete Category
         
     }
